documents_parser_llm/requirements.txt
langchain-chroma>=0.1
langchain-core>=0.2
uuid>=1
chromadb>=0.5
langchain_community.embeddings >=0.2
environ_config>=23
inquirer>=3.3
sentence_transformers>=3
PyMuPDF>=1.24
langchain_huggingface>=0.0